Split writing of EthernetFrame header attributes into NetworkByteBlock into single steps to have all attributes in the right byte order finally, changed size of EthernetDataPart to correct data type.

+++ src/kernel/network/internet/IP4Datagram.cpp
@@ -44,13 +44,16 @@ uint8_t IP4Datagram::copyDataTo(NetworkByteBlock *byteBlock) {
     if (this->ip4DataPart == nullptr || byteBlock == nullptr) {
         return 1;
     }
-    if (byteBlock->appendBytesInNetworkByteOrder(&this->header, sizeof(this->header))) {
+    if (byteBlock->appendBytesInNetworkByteOrder(
+            &this->header,
+            sizeof(this->header))
+            ) {
         return 1;
     }
     return this->ip4DataPart->copyDataTo(byteBlock);
 }
 
-uint16_t IP4Datagram::getLengthInBytes() {
+size_t IP4Datagram::getLengthInBytes() {
     return header.totalLength;
 }
