+++ src/kernel/network/internet/arp/ARPRequest.h
@@ -17,7 +17,7 @@ public:
 
     uint8_t copyDataTo(NetworkByteBlock *byteBlock) override;
 
-    uint16_t getLengthInBytes() override;
+    size_t getLengthInBytes() override;
 
     EtherType getEtherType() override;
 };
